+++ src/main/java/org/eclipse/microprofile/servicemesh/serviceb/ServiceBEndpoint.java
@@ -23,7 +23,7 @@
  *
  *******************************************************************************/
 
-package application.rest;
+package org.eclipse.microprofile.servicemesh.serviceb;
 
 
 import java.net.InetAddress;
