+++ src/main/webapp/WEB-INF/web.xml
@@ -2,5 +2,5 @@
 <web-app xmlns="http://xmlns.jcp.org/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd"
     version="3.1">
-    <display-name>ProducerService</display-name>
+    <display-name>Microprofile Servicemesh Sample Service B</display-name>
 </web-app>
